moleculesPropertiesPrediction/requirements.txt
#do not add here torch, torchvision and torchaudio because it causes the installation of libraries incompatible with rocm
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv
torch-geometric
numpy
pandas
scikit-learn
matplotlib
seaborn
psutil